==> notebooks/data_processing/github/Step 3 - CVE and Non-CVE GitHub Data Processing and Labeling.ipynb <==
# Processing GitHub Issues

import pandas as pd
import numpy as np
import os

gh_bq_issues = pd.read_csv('GH_issues.csv')
gh_bq_issues.info()

gh_bq_issues[pd.isnull(gh_bq_issues.issue_id)].shape

gh_bq_issues = gh_bq_issues[~pd.isnull(gh_bq_issues.issue_id)]
gh_bq_issues.shape

gh_bq_issues.issue_id.value_counts().iloc[:10]

# Checking and Removing Issues with wrong timestamp

[item for item in gh_bq_issues.issue_created_at.tolist() if item.startswith('3')]

[item for item in gh_bq_issues.issue_updated_at.tolist() if item.startswith('3')]

[item for item in gh_bq_issues.issue_closed_at.tolist() if not pd.isnull(item) and item.startswith('3')]

gh_bq_issues = gh_bq_issues[~(gh_bq_issues.issue_created_at == '3012-03-02T14:10:15Z')]

gh_bq_issues.issue_created_at = pd.to_datetime(gh_bq_issues.issue_created_at)
gh_bq_issues.issue_updated_at = pd.to_datetime(gh_bq_issues.issue_updated_at)
gh_bq_issues.issue_closed_at = pd.to_datetime(gh_bq_issues.issue_closed_at)
gh_bq_issues.shape

# Removing Duplicate Issues 
### Keeping each issue record based on last updated timestamp

gh_bq_issues = gh_bq_issues.loc[gh_bq_issues.groupby('issue_id').issue_updated_at.idxmax(skipna=False)]
gh_bq_issues.shape

gh_bq_issues.issue_id.value_counts().iloc[:10]

# Reading in Positive CVE related GitHub Issue Links

gh_pos_issue_links = pd.read_csv('gh_sec_issue_links.csv')
gh_pos_issue_links.info()

cve_issue_links = gh_pos_issue_links.issue.tolist()
len(cve_issue_links)

# Extracting CVE issues from the larger issues dataset

cve_issues = gh_bq_issues[gh_bq_issues.issue_url.isin(cve_issue_links)]
cve_issues.shape

cve_issues.issue_url.value_counts().iloc[:10]

cve_issues.head()

found_issue_urls = cve_issues.issue_url.tolist()

# Getting CVE Positive issues not found in Big Query Data

not_found_issue_urls = list(set(cve_issue_links) - set(found_issue_urls))
not_found_issue_urls[:10], len(not_found_issue_urls)

# Extracting not found CVE issues with GitHub API

import requests
from tqdm import tqdm
import re

not_found_api_urls = ['https://api.github.com/repos/'+re.search(r'.*github.com/(.*)', link, re.I).groups()[0]
                          for link in not_found_issue_urls]
not_found_api_urls[:10]

data = []
for link in tqdm(not_found_api_urls):
    response = requests.get(link,
                            auth=('dipanjanS', os.environ['GITHUB_TOKEN']))
    if not response.status_code == 200:
        print('Failed for link: '+link)
        # log this later
    else:
        content = response.json()
        issue_dict = {
            'repo_name': re.search(r'.*github.com/repos/(.*?)/issues', 
                                   link, re.I).groups()[0],
            'event_type': 'IssuesEvent',
            'actor_id': content.get('user').get('id'),
            'actor_name': content.get('user').get('login'),
            'issue_status': content.get('state'),
            'issue_api_url': content.get('url'),
            'issue_url': content.get('html_url'),
            'issue_creator_name': content.get('user').get('login'),
            'issue_creator_api_url': content.get('user').get('url'),
            'issue_creator_url': content.get('user').get('html_url'),
            'comment_count': content.get('comments'),
            'issue_id': content.get('id'),
            'issue_number': content.get('number'),
            'issue_created_at': content.get('created_at'),
            'issue_updated_at': content.get('updated_at'),
            'issue_closed_at': content.get('closed_at'),
            'issue_title': content.get('title'),
            'issue_body': content.get('body')
        }
        if issue_dict:
            data.append(issue_dict)

print(len(data))

cve_issues_nf = pd.DataFrame(data)
cve_issues_nf = cve_issues_nf[cve_issues.columns.tolist()]
cve_issues_nf.info()

cve_issues_nf.head()

# Labeling CVE and non-CVE GitHub issues 

gh_bq_issues.shape

gh_bq_issues_positive = pd.concat([cve_issues, cve_issues_nf], axis=0)
gh_bq_issues_negative = gh_bq_issues.drop(cve_issues.index.tolist())
gh_bq_issues_positive.shape, gh_bq_issues_negative.shape

gh_bq_issues_positive['class_label'] = 2
gh_bq_issues_negative['class_label'] = 0

gh_bq_issues_processed = pd.concat([gh_bq_issues_negative, gh_bq_issues_positive], axis=0)
gh_bq_issues_processed.shape

gh_bq_issues_processed.tail()

gh_bq_issues_processed.to_csv('GH_issues_clean_labeled.csv', index=False)

# Processing Pull Requests

gh_bq_prs = pd.read_csv('GH_pull_requests.csv')
gh_bq_prs.info()

# Adding in missing GitHub Pull Request Links

gh_bq_prs = gh_bq_prs[~pd.isnull(gh_bq_prs.pr_title)]

def fill_missing_links(record):
    repo_name = record['repo_name']
    pr_number = str(record['pr_number'])
    if pd.isnull(record['pr_api_url']):
        record['pr_api_url'] = 'https://api.github.com/repos/'+repo_name+'/pulls/'+pr_number
    if pd.isnull(record['pr_url']):
        record['pr_url'] = 'https://github.com/'+repo_name+'/pull/'+pr_number 
    if pd.isnull(record['pr_diff_url']):
        record['pr_diff_url'] = 'https://github.com/'+repo_name+'/pull/'+pr_number+'.diff' 
    if pd.isnull(record['pr_patch_url']):
        record['pr_patch_url'] = 'https://github.com/'+repo_name+'/pull/'+pr_number+'.patch'
    return record

gh_bq_prs = gh_bq_prs.apply(lambda row: fill_missing_links(row), axis=1)
gh_bq_prs.info()

# Checking and Removing Pull Requests with wrong timestamp

[item for item in gh_bq_prs.pr_created_at.tolist() if not pd.isnull(item) and item.startswith('3')]

[item for item in gh_bq_prs.pr_updated_at.tolist() if not pd.isnull(item) and item.startswith('3')]

[item for item in gh_bq_prs.pr_closed_at.tolist() if not pd.isnull(item) and item.startswith('3')]

[item for item in gh_bq_prs.pr_merged_at.tolist() if not pd.isnull(item) and item.startswith('3')]

# Removing Duplicate Pull Requests 
### Keeping each PR record based on last row where timestamp is missing

gh_bq_prs_missing_info = gh_bq_prs[pd.isnull(gh_bq_prs.pr_updated_at)]
gh_bq_prs_full_info = gh_bq_prs[~pd.isnull(gh_bq_prs.pr_updated_at)]
gh_bq_prs_missing_info.shape, gh_bq_prs_full_info.shape

gh_bq_prs_missing_info = gh_bq_prs_missing_info.drop_duplicates(subset=['pr_url'], keep="last")
gh_bq_prs_missing_info.shape

# Removing Duplicate Pull Requests 
### Keeping each PR record based on last updated timestamp

gh_bq_prs_full_info.pr_created_at = pd.to_datetime(gh_bq_prs_full_info.pr_created_at)
gh_bq_prs_full_info.pr_updated_at = pd.to_datetime(gh_bq_prs_full_info.pr_updated_at)
gh_bq_prs_full_info.pr_closed_at = pd.to_datetime(gh_bq_prs_full_info.pr_closed_at)
gh_bq_prs_full_info.pr_merged_at = pd.to_datetime(gh_bq_prs_full_info.pr_merged_at)
gh_bq_prs_full_info.shape

gh_bq_prs_full_info.info()

gh_bq_prs_full_info = gh_bq_prs_full_info.loc[gh_bq_prs_full_info.groupby('pr_url').pr_updated_at.idxmax(skipna=False)]
gh_bq_prs_full_info.shape

gh_bq_prs = pd.concat([gh_bq_prs_full_info, gh_bq_prs_missing_info], axis=0)
gh_bq_prs.shape

# Reading in Positive CVE related GitHub Pull Request Links

gh_pos_pr_links = pd.read_csv('gh_sec_pr_links.csv')
gh_pos_pr_links.info()

cve_pr_links = gh_pos_pr_links.pull_request.tolist()
len(cve_pr_links)

# Extracting CVE Pull Requests from the larger PRs dataset

cve_prs = gh_bq_prs[gh_bq_prs.pr_url.isin(cve_pr_links)]
cve_prs.shape

cve_prs.head()

# Getting CVE Positive Pull Requests not found in Big Query Data

found_pr_urls = cve_prs.pr_url.tolist()
not_found_pr_urls = list(set(cve_pr_links) - set(found_pr_urls))
not_found_pr_urls[:10], len(not_found_pr_urls)

not_found_api_urls = [re.sub('/pull/', '/pulls/', 
                             ('https://api.github.com/repos/'+re.search(r'.*github.com/(.*)', link, re.I)
                              .groups()[0]), re.I)
                          for link in not_found_pr_urls]
not_found_api_urls[:10]

# Extracting not found CVE Pull Requests with GitHub API

data = []
for link in tqdm(not_found_api_urls):
    response = requests.get(link,
                            auth=('dipanjanS', os.environ['GITHUB_TOKEN']))
    if not response.status_code == 200:
        print('Failed for link: '+link)
        # log this later
    else:
        content = response.json()
        pr_dict = {
            'repo_name': re.search(r'.*github.com/repos/(.*?)/pulls', 
                                   link, re.I).groups()[0],
            'event_type': 'PullRequestEvent',
            'actor_id': content.get('user').get('id'),
            'actor_name': content.get('user').get('login'),
            'pr_status': content.get('state'),
            'pr_id': content.get('id'),
            'pr_number': content.get('number'),
            'pr_api_url': content.get('url'),
            'pr_url': content.get('html_url'),
            'pr_diff_url': content.get('diff_url'),
            'pr_patch_url': content.get('patch_url'),
            'pr_creator_name': content.get('user').get('login'),
            'pr_creator_api_url': content.get('user').get('url'),
            'pr_creator_url': content.get('user').get('html_url'),         
            'pr_created_at': content.get('created_at'),
            'pr_updated_at': content.get('updated_at'),
            'pr_closed_at': content.get('closed_at'),
            'pr_merged_at': content.get('merged_at'),
            'pr_merged_status': content.get('merged'),
            'pr_comments_count': content.get('comments'),
            'pr_review_comments_count': content.get('review_comments'),
            'pr_commits_count': content.get('commits'),
            'pr_additions_count': content.get('additions'),
            'pr_deletions_count': content.get('deletions'),
            'pr_changed_files_count': content.get('changed_files'),
            'pr_title': content.get('title'),
            'pr_body': content.get('body')
        }
        if pr_dict:
            data.append(pr_dict)

print(len(data))

cve_prs_nf = pd.DataFrame(data)
cve_prs_nf = cve_prs_nf[cve_prs.columns.tolist()]
cve_prs_nf.info()

cve_prs_nf.head()

# Labeling CVE and non-CVE GitHub Pull Requests 

gh_bq_prs.shape

gh_bq_prs_positive = pd.concat([cve_prs, cve_prs_nf], axis=0)
gh_bq_prs_negative = gh_bq_prs.drop(cve_prs.index.tolist())
gh_bq_prs_positive.shape, gh_bq_prs_negative.shape

gh_bq_prs_positive['class_label'] = 2
gh_bq_prs_negative['class_label'] = 0

gh_bq_prs_processed = pd.concat([gh_bq_prs_negative, gh_bq_prs_positive], axis=0)
gh_bq_prs_processed.shape

gh_bq_prs_processed.to_csv('GH_prs_clean_labeled.csv', index=False)
