==> notebooks/data_processing/github/Step 4 - Labeling GitHub Data for Security Vulnerabilities - Regular Expressions.ipynb <==
import re
import pandas as pd
import numpy as np

issues_df = pd.read_csv('GH_issues_clean_labeled.csv')
issues_df.info()

issues_df['issue_title'] = issues_df.issue_title.fillna('')
issues_df['issue_body'] = issues_df.issue_body.fillna('')
issues_df['description'] = issues_df['issue_title'].map(str) + ' ' + issues_df['issue_body']

issues_df.class_label.value_counts()

prs_df = pd.read_csv('GH_prs_clean_labeled.csv')
prs_df.info()

prs_df['pr_title'] = prs_df.pr_title.fillna('')
prs_df['pr_body'] = prs_df.pr_body.fillna('')
prs_df['description'] = prs_df['pr_title'].map(str) + ' ' + prs_df['pr_body']

prs_df.class_label.value_counts()

df = pd.concat([issues_df[['description', 'class_label']], 
                prs_df[['description', 'class_label']]], axis=0).reset_index(drop=True)
df.info()

df.head()

df.class_label.value_counts()

def security_vulnerability_labeler(issue_description):
    strong_vuln_pattern = ('(?i)(advisory|attack|(un)?authoriz(e|ation)|'
                           'clickjack|cross.site|csrf|\\bCVE.*?\\b|deadlock|'
                           'denial.of.service|\\bEOP\\b|exploit|hijack|'
                           'infinite.loop|malicious|\\bNVD\\b|OSVDB|'
                           '\\bRCE\\b|\\bReDoS\\b|\\bDDoS\\b|remote.code.execution|'
                           'security|victim|\\bvuln|\\bXEE\\b|\\bXSRF\\b|'
                           '\\bXSS\\b|\\bXXE\\b)')

    medium_vuln_pattern = ('(?i)(authenticat(e|ion)|brute force|bypass|'
                           'constant.time|crack|credential|\\bDoS\\b|'
                           'expos(e|ing)|hack|harden|injection|lockout|'
                           'overflow|password|\\bPoC\\b|proof.of.concept|'
                           'poison|priv(ilege|elege|elage|lage)|\\b(in)?secur(e|ity)|'
                           '(de)?serializ|spoof|timing|traversal)')

    low_vuln_pattern = ('(?i)(abuse|compliant|constant.time|credential|\\bcrypto|'
                        'escalate|exhaustion|forced|infinite|RFC\\d{4,5})')
    
    if (re.findall(low_vuln_pattern, issue_description, re.I) or
        re.findall(medium_vuln_pattern, issue_description, re.I) or
        re.findall(strong_vuln_pattern, issue_description, re.I)):
        return 1
    else:
        return 0

negative_df = df[df.class_label == 0]
positive_df = df[df.class_label == 2]

negative_df['label'] = negative_df.apply(lambda row: security_vulnerability_labeler(row.description), axis=1)

positive_df['label'] = positive_df['class_label']

df = pd.concat([positive_df[['description', 'label']],
                negative_df[['description', 'label']]], axis=0).reset_index(drop=True)
df.info()

df['label'].value_counts()

df.to_csv('GH_processed_labeled_issues_prs.csv', index=False)
