==> notebooks/data_processing/github/Step 5 - Data Pre-processing.ipynb <==
import pandas as pd
import numpy as np
import re
import dill
import tqdm
import gc

df = pd.read_csv('GH_processed_labeled_issues_prs.csv')
df.head()

from sklearn.model_selection import train_test_split

X = df.description.values
y = df.label.values

X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.2, random_state=42)
X_train.shape, X_test.shape, y_train.shape, y_test.shape

from collections import Counter
print('Train:', Counter(y_train))
print('Test:', Counter(y_test))

import contractions
from bs4 import BeautifulSoup
import unicodedata
import re


def strip_html_tags(text):
    soup = BeautifulSoup(text, "lxml")
    [s.extract() for s in soup(['iframe', 'script'])]
    stripped_text = soup.get_text()
    stripped_text = re.sub(r'[\r|\n|\r\n]+', '\n', stripped_text, re.I)
    return stripped_text


def remove_urls(text):
    url_pattern = '((https?:\/\/)(\s)*(www\.)?|(www\.))(\s)*((\w|\s)+\.)*([\w\-\s]+\/)*([\w\-]+)((\?)?[\w\s]*=\s*[\w\%&]*)*'
    text = re.sub(url_pattern, ' ', text, re.I)
    return text

def remove_checklists(text):
    checklist_pattern = r'\[[xX\.\s]\]'
    text = re.sub(checklist_pattern, ' ', text, re.I|re.DOTALL)
    return text

def remove_accented_chars(text):
    text = unicodedata.normalize('NFKD', text).encode('ascii', 'ignore').decode('utf-8', 'ignore')
    return text


def expand_contractions(text):
    return contractions.fix(text)



def remove_special_characters(text, remove_digits=False):
    pattern = r'[^a-zA-Z0-9\s]' if not remove_digits else r'[^a-zA-Z\s]'
    text = re.sub(pattern, ' ', text)
    return text


def pre_process_document(document):
    
    # strip HTML
    document = strip_html_tags(document)
    
    # remove URLS
    document = remove_urls(document)
    
    # remove checklists
    document = remove_checklists(document)
    
    # expand contractions    
    document = expand_contractions(document)
    
    # lower case
    document = document.lower()
    
    # remove extra newlines (often might be present in really noisy text)
    document = document.translate(document.maketrans("\n\t\r", "   "))
    
    # remove accented characters
    document = remove_accented_chars(document)
    
    
               
    # remove special characters and\or digits    
    # insert spaces between special characters to isolate them    
    special_char_pattern = re.compile(r'([{.(-)!}])')
    document = special_char_pattern.sub(" \\1 ", document)
    document = remove_special_characters(document, remove_digits=False)  
    
    # remove only numbers
    document = re.sub(r'\b\d+\b', ' ', document)
        
    # remove extra whitespace
    document = re.sub(' +', ' ', document)
    document = document.strip()
    
    return document


pre_process_corpus = np.vectorize(pre_process_document)

gc.collect()

docs = []
for idx, record in tqdm.tqdm(enumerate(X_train)):
    doc = pre_process_document(record)
    docs.append(doc)
    if ((idx+1) % 50000 == 0) or ((idx+1) == len(X_train)):
        with open('Xtrain_norm.pkl', 'ab+') as f:
            dill.dump(docs, f)
        print('Processed and stored {} docs'.format(len(docs)))
        docs = []

gc.collect()

docs = []
for idx, record in tqdm.tqdm(enumerate(X_test)):
    doc = pre_process_document(record)
    docs.append(doc)
    if ((idx+1) % 50000 == 0) or ((idx+1) == len(X_test)):
        with open('Xtest_norm.pkl', 'ab+') as f:
            dill.dump(docs, f)
        print('Processed and stored {} docs'.format(len(docs)))
        docs = []

gc.collect()

with open('ytrain_labels.pkl', 'wb') as f:
    dill.dump(y_train, f)
with open('ytest_labels.pkl', 'wb') as f:
    dill.dump(y_test, f)

gc.collect()
